Split useMutableSource tests into separate suite
